created day-night-theme reusable component
I created a toggle-button element and added css to display it like a toggle button, i added it to navbar, and added js for changing background color if button is toggled. I imported different css of components which background I want to be changed if btn toggled. i removed previous day-theme classNames and I added .dark classname for elements which are supposed to change backgroud